Reflect CloudAI team's comments
