+++ pyproject.toml
@@ -26,7 +26,6 @@ dependencies = [
   "pydantic==2.8.2",
   "jinja2==3.1.6",
   "websockets==15.0.1",
-  "elasticsearch==7.12.1",
 ]
   [project.scripts]
   cloudai = "cloudai.__main__:main"
