--- src/cloudai/data/model/gsw_context.py
@@ -1,32 +1,0 @@
-# SPDX-FileCopyrightText: NVIDIA CORPORATION & AFFILIATES
-# Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from pydantic import BaseModel, Field
-
-
-class GSWContext(BaseModel):
-    """
-    Hold GSW metadata common to all workloads.
-
-    Attributes:
-      gsw_version: The version of the GSW.
-    """
-
-    gsw_version: str = Field(alias="s_gsw_version")
-
-    model_config = {
-        "populate_by_name": True,
-    }
