--- src/cloudai/data/repository/base.py
@@ -1,37 +1,0 @@
-# SPDX-FileCopyrightText: NVIDIA CORPORATION & AFFILIATES
-# Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from abc import ABC, abstractmethod
-from typing import Any, Dict
-
-
-class BaseDataRepository(ABC):
-    """Abstract interface for storing, retrieving, and searching data entries."""
-
-    @abstractmethod
-    def store(self, entry: Dict[str, Any]) -> None:
-        """Store a data entry."""
-        pass
-
-    @abstractmethod
-    def retrieve(self, identifier: str) -> Dict[str, Any]:
-        """Retrieve a data entry by its identifier."""
-        pass
-
-    @abstractmethod
-    def search(self, query: Dict[str, Any]) -> Dict[str, Any]:
-        """Execute a search query and return the results as a dictionary."""
-        pass
