+++ src/cloudai/workloads/nemo_run/__init__.py
@@ -15,6 +15,7 @@
 # limitations under the License.
 
 
+from .llmb_report_generation_strategy import NeMoRunLLMBReportGenerationStrategy
 from .nemo_run import Data, Log, LogCkpt, NeMoRunCmdArgs, NeMoRunTestDefinition, Trainer, TrainerStrategy
 from .report_generation_strategy import NeMoRunReportGenerationStrategy
 from .slurm_command_gen_strategy import NeMoRunSlurmCommandGenStrategy
@@ -24,6 +25,7 @@ __all__ = [
     "Log",
     "LogCkpt",
     "NeMoRunCmdArgs",
+    "NeMoRunLLMBReportGenerationStrategy",
     "NeMoRunReportGenerationStrategy",
     "NeMoRunSlurmCommandGenStrategy",
     "NeMoRunTestDefinition",
