+++ src/cloudai/workloads/nemo_run/data/http_data_repository.py
@@ -19,41 +19,25 @@ from typing import Any, Dict
 
 import requests
 
-from .base import BaseDataRepository
 
-
-class HttpDataRepository(BaseDataRepository):
+class HttpDataRepository:
     """Data repository using HTTP API (requests) instead of elastic client."""
 
     def __init__(
         self,
+        post_endpoint: str,
         token: str,
         index: str,
-        base_url_data: str,
-        base_url_search: str,
         verify_certs: bool = True,
     ) -> None:
+        self.post_endpoint = post_endpoint
         self.token = token
         self.index = index
-        self.base_url_data = base_url_data
-        self.base_url_search = base_url_search
         self.verify = verify_certs
 
     def store(self, entry: Dict[str, Any]) -> None:
-        endpoint = f"{self.base_url_data}/{self.index}/posting"
+        endpoint = f"{self.post_endpoint}/{self.index}/posting"
         headers = {"Authorization": f"Bearer {self.token}", "Content-Type": "application/json"}
         payload = json.dumps([entry])
         response = requests.post(endpoint, data=payload, headers=headers, verify=self.verify)
         response.raise_for_status()
-
-    def retrieve(self, identifier: str) -> Dict[str, Any]:
-        query = {"query": {"match": {"_id": identifier}}}
-        results = self.search(query)
-        return results.get("hits", {}).get("hits", [{}])[0].get("_source", {})
-
-    def search(self, query: Dict[str, Any]) -> Dict[str, Any]:
-        endpoint = f"{self.base_url_search}/res/v1/es/{self.index}/_search"
-        headers = {"Authorization": f"Bearer {self.token}", "Content-Type": "application/json"}
-        response = requests.post(endpoint, headers=headers, data=json.dumps(query), verify=self.verify)
-        response.raise_for_status()
-        return response.json()
